* Rename test to match with name of file being tested. * Fix existing tests to verify parameters on all branches -- the reason the tests were passing   was that it was taking the default branch in the test cases * Update all of the tests to directly call run rather than route everything through ray.
